requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2

# streamlit run app.py
# to get api key
# https://makersuite.google.com/app/apikey